accepting data and display
Complete Home page and Product info page

+++ lib/Models/product_model.dart
@@ -32,7 +32,7 @@ class ProductModel {
 }
 
 class RatingModel {
-  final Float rate;
+  final double rate;
   final int count;
   RatingModel({required this.rate, required this.count});
